pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
dnspython = "^2.4.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"

[tool.autopep8]
max_line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
